apps/backend/requirements.txt
fastapi
uvicorn
supabase
psycopg2-binary
google-generativeai
python-multipart
pymupdf
pdfplumber
pytesseract
pillow
sentence-transformers
celery
redis
pydantic
python-dotenv
loguru
